nmstate: build rpm under en_US LC_TIME

+++ prepare/nmstate.sh
@@ -5,7 +5,13 @@ rm -rf /usr/bin/python && ln -s /usr/bin/python3 /usr/bin/python
 git clone https://github.com/nmstate/nmstate
 cd nmstate
 git checkout $(git tag |tail -1)
-sh packaging/make_rpm.sh
-rm -rf nmstate-*.src.rpm
-yum -y localinstall python3-libnmstate* nmstate-*
-python -m pip install pytest
+LC_TIME=en_US-UTF-8 sh packaging/make_rpm.sh; rc=$?
+if test $rc -eq 0; then
+    rm -rf nmstate-*.src.rpm
+    yum -y localinstall python3-libnmstate* nmstate-*
+    python -m pip install pytest
+else
+    printf "\n\n* COMPILATION FAILED!\n\n"
+fi
+
+exit $rc
